Module aws_sdk_fis::error

Modules

conflict_exception

See ConflictException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

validation_exception

See ValidationException

Structs

ConflictException

The request could not be processed because of a conflict.

CreateExperimentTemplateError

DeleteExperimentTemplateError

GetActionError

GetExperimentError

GetExperimentTemplateError

ListActionsError

ListExperimentTemplatesError

ListExperimentsError

ListTagsForResourceError

ResourceNotFoundException

The specified resource cannot be found.

ServiceQuotaExceededException

You have exceeded your service quota.

StartExperimentError

StopExperimentError

TagResourceError

UntagResourceError

UpdateExperimentTemplateError

ValidationException

The specified input is not valid, or fails to satisfy the constraints for the request.

Enums

CreateExperimentTemplateErrorKind

DeleteExperimentTemplateErrorKind

GetActionErrorKind

GetExperimentErrorKind

GetExperimentTemplateErrorKind

ListActionsErrorKind

ListExperimentTemplatesErrorKind

ListExperimentsErrorKind

ListTagsForResourceErrorKind

StartExperimentErrorKind

StopExperimentErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateExperimentTemplateErrorKind