Module aws_sdk_fis::model

Modules

action

See Action

action_parameter

See ActionParameter

action_summary

See ActionSummary

action_target

See ActionTarget

create_experiment_template_action_input

See CreateExperimentTemplateActionInput

create_experiment_template_stop_condition_input

See CreateExperimentTemplateStopConditionInput

create_experiment_template_target_input

See CreateExperimentTemplateTargetInput

experiment

See Experiment

experiment_action

See ExperimentAction

experiment_action_state

See ExperimentActionState

experiment_state

See ExperimentState

experiment_stop_condition

See ExperimentStopCondition

experiment_summary

See ExperimentSummary

experiment_target

See ExperimentTarget

experiment_target_filter

See ExperimentTargetFilter

experiment_template

See ExperimentTemplate

experiment_template_action

See ExperimentTemplateAction

experiment_template_stop_condition

See ExperimentTemplateStopCondition

experiment_template_summary

See ExperimentTemplateSummary

experiment_template_target

See ExperimentTemplateTarget

experiment_template_target_filter

See ExperimentTemplateTargetFilter

experiment_template_target_input_filter

See ExperimentTemplateTargetInputFilter

update_experiment_template_action_input_item

See UpdateExperimentTemplateActionInputItem

update_experiment_template_stop_condition_input

See UpdateExperimentTemplateStopConditionInput

update_experiment_template_target_input

See UpdateExperimentTemplateTargetInput

Structs

Action

Describes an action. For more information, see AWS FIS actions

ActionParameter

Describes a parameter for an action.

ActionSummary

Provides a summary of an action.

ActionTarget

Describes a target for an action.

CreateExperimentTemplateActionInput

Specifies an action for an experiment template.

CreateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment template.

CreateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or

Experiment

Describes an experiment.

ExperimentAction

Describes the action for an experiment.

ExperimentActionState

Describes the state of an action.

ExperimentState

Describes the state of an experiment.

ExperimentStopCondition

Describes the stop condition for an experiment.

ExperimentSummary

Provides a summary of an experiment.

ExperimentTarget

Describes a target for an experiment.

ExperimentTargetFilter

Describes a filter used for the target resources in an experiment.

ExperimentTemplate

Describes an experiment template.

ExperimentTemplateAction

Describes an action for an experiment template.

ExperimentTemplateStopCondition

Describes a stop condition for an experiment template.

ExperimentTemplateSummary

Provides a summary of an experiment template.

ExperimentTemplateTarget

Describes a target for an experiment template.

ExperimentTemplateTargetFilter

Describes a filter used for the target resources in an experiment template.

ExperimentTemplateTargetInputFilter

Describes a filter used for the target resource input in an experiment template.

UpdateExperimentTemplateActionInputItem

Specifies an action for an experiment template.

UpdateExperimentTemplateStopConditionInput

Specifies a stop condition for an experiment. You can define a stop condition as a CloudWatch alarm.

UpdateExperimentTemplateTargetInput

Specifies a target for an experiment. You must specify at least one Amazon Resource Name (ARN) or at least one resource tag. You cannot specify both.

Enums

ExperimentActionStatus

ExperimentStatus